+++ no-style/README.md
@@ -122,3 +122,51 @@ npm config set prefix /usr/local
 ```
 
 로 글로벌 설치 경로를 변경해 주었다.
+
+
+
+
+
+성능 최적화 이전
+![Alt text](image-1.png)
+
+![Alt text](image-2.png)
+
+
+성능 최적화 이후
+
+![Alt text](image-3.png)
+
+성능 최적화 이후의 변화는 렌더링 되는 데이터양이 많지 않아서 정확히 식별이 불가함
+
+
+10000개의 벌크 데이터를 db에 기록한 뒤에 10000개를 렌더링 하는것에 대해서 최적화 이전과 이후를 비교한다.
+
+
+벌크 데이터를 이용한 최적화 이전 성능 측정
+
+![Alt text](image-4.png)
+
+
+벌크 데이터를 이용한 최적화 이후 성능 측정
+
+![Alt text](image-5.png)
+
+
+
+렌더링 시간이 대폭 줄어든 것을 확인할 수 있다.
+
+
+서버 성능 테스트
+
+입력 데이터
+![Alt text](image-7.png)
+
+
+1개의 스레드를 사용하여 1개의 커넥트를 1초간 유지하여 백엔드 성능 측정
+![Alt text](image-8.png)
+
+
+10개의 스레드를 생성하여 100개의 커넥트를 30초간 유지하여 백엔드 성능 측정
+
+![Alt text](image-9.png)
