+++ script.lua
@@ -1,0 +1,3 @@
+wrk.method = "POST"
+wrk.body = "{ "keyword": "캡스톤", "type": "subject" }"
+wrk.headers["Content-Type"] = "application/json"
